Move util files and functions
