+++ src/components/Admin/AccountManagement/ChangeName.tsx
@@ -19,13 +19,10 @@ import {
   FormHelperText,
 } from '@chakra-ui/react';
 import {CheckCircleIcon} from '@chakra-ui/icons';
-import {
-  handleReauthenticationWithPassword,
-  PasswordFormInput,
-  SubmitButton,
-} from '../Util';
+import {PasswordFormInput, SubmitButton} from '../ComponentUtil';
 import {firestore, firebaseAuth} from '../../../firebase';
 import {useTranslation} from 'react-i18next';
+import {handleReauthenticationWithPassword} from '../Authentication/Util';
 
 /**
  * Props for ChangeNameModal component. Used for type safety.
