+++ src/components/Admin/Authentication/SignUp.tsx
@@ -2,13 +2,13 @@ import React, {useState} from 'react';
 import {Text, Heading, Link, Divider} from '@chakra-ui/react';
 import {
   SubmitButton,
-  signInWithGoogle,
   EmailFormInput,
   PasswordFormInput,
-} from '../Util';
+} from '../ComponentUtil';
 import {firebaseAuth} from '../../../firebase';
 import {UnauthenticatedPageProps} from '../UnauthenticatedAdmin';
 import {useTranslation} from 'react-i18next';
+import {signInWithGoogle} from './Util';
 
 /**
  * Component to sign up.
