+++ src/components/Admin/ComponentUtil.tsx
@@ -11,9 +11,7 @@ import {
   Box,
   FormHelperText,
 } from '@chakra-ui/react';
-import firebase, {firebaseAuth} from '../../firebase';
 import {useTranslation} from 'react-i18next';
-import {TFunction} from 'i18next';
 
 /**
  * Props for PasswordVisibilityToggle component. Used for type safety.
@@ -214,133 +212,4 @@ const SubmitButton: ({
   );
 };
 
-/**
- * Signs in or signs up a user with Google authentication through a pop up.
- * @param event - submit form event
- * @param setError - function to set error state for any errors from Google
- * @param setIsLoading - function to set loading state
- * @param translate - function to translate text using i18n-next
- */
-async function signInWithGoogle(
-  event: React.FormEvent<HTMLFormElement>,
-  setError: React.Dispatch<React.SetStateAction<string>>,
-  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>,
-  translate: TFunction
-): Promise<void> {
-  // Prevents submission before sign in is complete
-  event.preventDefault();
-
-  setIsLoading(true);
-  try {
-    await firebaseAuth.signInWithPopup(new firebase.auth.GoogleAuthProvider());
-  } catch (error) {
-    // Error codes from Firebase documentation
-    switch (error.code) {
-      case 'auth/account-exists-with-different-credential': {
-        firebaseAuth
-          .fetchSignInMethodsForEmail(error.email)
-          .then(signInMethodArray => {
-            const errorMessageTemplate = translate('wrongMethod');
-            const methods = signInMethodArray.join(', ');
-            setError(errorMessageTemplate + methods);
-          })
-          .catch(error => {
-            switch (error.code) {
-              case 'auth/invalid-email': {
-                setError(translate('invalidEmail'));
-                break;
-              }
-              default: {
-                setError(translate('tryAgain'));
-                break;
-              }
-            }
-          });
-        break;
-      }
-      case 'auth/cancelled-popup-request': {
-        // No error
-        break;
-      }
-      case 'auth/popup-closed-by-user': {
-        // No error
-        break;
-      }
-      case 'auth/popup-blocked': {
-        setError(translate('popUpsBlocked'));
-        break;
-      }
-      default: {
-        setError(translate('common:error'));
-        break;
-      }
-    }
-    setIsLoading(false);
-  }
-}
-
-/**
- * Handles reauthentication of a password-based user before account update operations.
- * @returns error message or empty string if no error
- *
- * @throws No user
- * Thrown if the currentUser is null
- *
- * @throws No email
- * Thrown if the currentUser's email is null
- */
-async function handleReauthenticationWithPassword(
-  password: string,
-  translate: TFunction
-): Promise<string> {
-  if (!firebaseAuth.currentUser) throw new Error(translate('noUser'));
-  if (!firebaseAuth.currentUser.email) throw new Error(translate('noEmail'));
-  const credential = firebase.auth.EmailAuthProvider.credential(
-    firebaseAuth.currentUser.email,
-    password
-  );
-
-  try {
-    await firebaseAuth.currentUser.reauthenticateWithCredential(credential);
-    return '';
-  } catch (error) {
-    // Error codes from Firebase documentation
-    // For fatal errors, user will be signed out and redirected to sign in
-    const fatalErrors = [
-      'auth/user-mismatch',
-      'auth/user-not-found',
-      'auth/invalid-credential',
-      'auth/invalid-email',
-    ];
-    if (fatalErrors.includes(error.code)) firebaseAuth.signOut();
-
-    switch (error.code) {
-      case 'auth/wrong-password': {
-        return translate('incorrectPassword');
-      }
-      default: {
-        return translate('unknownError') + error.message;
-      }
-    }
-  }
-}
-
-/**
- * Interface to describe the contents of a user's doc in Firestore
- */
-interface User {
-  email: string;
-  name: string;
-  admin: boolean;
-  userId: string;
-}
-
-export {
-  handleReauthenticationWithPassword,
-  SubmitButton,
-  signInWithGoogle,
-  EmailFormInput,
-  PasswordFormInput,
-};
-
-export type {User};
+export {SubmitButton, EmailFormInput, PasswordFormInput};
