+++ src/util.ts
@@ -1,15 +1,4 @@
 /**
- *
- * @param value - the value being checked for validity, should be string or boolean
- * @param type - the valid type for the value being checked
- *
- * @returns true if the inputted value is not undefined and is of the type `type`
- */
-function validData(value: string | boolean, type: string): boolean {
-  return value !== undefined && typeof value === type;
-}
-
-/**
  * Upper bounds on AQI categories.
  *
  * Source: https://www.airnow.gov/aqi/aqi-basics/
@@ -23,4 +12,4 @@ const aqiCutoffs = {
   veryUnhealthy: 300, // Very unhealthy for all individuals (201-300)
 };
 
-export {validData, aqiCutoffs};
+export {aqiCutoffs};
